cyberbridge_backend/requirements.txt
fastapi[standard]==0.133.0
requests==2.32.5
httpx==0.28.1
SQLAlchemy==2.0.47
python-dotenv==1.2.1
pydantic==2.12.5
psycopg2-binary==2.9.11
email-validator==2.3.0
passlib==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
mammoth==1.11.0
alembic==1.18.4
python-dateutil==2.9.0.post0
PyPDF2==3.0.1
APScheduler==3.11.2
pandas==3.0.1
openpyxl==3.1.5
pyotp==2.9.0
qrcode[pil]==8.2
reportlab==4.4.10
Pillow==12.1.1
beautifulsoup4==4.14.3
Authlib==1.6.8
